Add OAuth support for Pywikibot
This change depends on mwoauth for OAuth support. A new OAuthLoginManager is
used for OAuth tokens retrieval. The whole retrieval process is added to
scripts.login module. OAuth tokens are set in config.authenticate and
enable automatically if applicable.

Tests are configured by setting OAuth related environment variables. Family
names of orain test and beta wiki are changed to adapted to OAuth testing.

Bug: T102602
Change-Id: Ib3b464df3f5f8607dc84ff2d5f2969de4c693561